refactor(auth, user): move endpoints for user existance check and show profile options from auth to user domain
